Move new test into separate file.
Otherwise pytest hangs on the CI with MacOS and Python 3.10.
